move progress callback to core, default enable on None
