+++ nobodywho/python/Cargo.toml
@@ -22,7 +22,6 @@ pyo3-introspection = "0.28.3"
 pyo3-log = "0.13.2"
 pythonize = "0.28.0"
 nom = "8.0.0"
-indicatif = "0.18"
 
 
 [dependencies.pyo3]
